fix: additional ParamValue, enum, and bit-field formatting fixes
- mavlink_parameter_server.cpp: use .get_string() for ParamValue args
  and static_cast<int> for Result enum
- system_tests: cast msg.msgid bit-field and Mission::Result/Camera::Mode
  enums to int for std::format compatibility

+++ cpp/src/system_tests/ftp_download_file.cpp
@@ -270,7 +270,7 @@ TEST(SystemTest, FtpDownloadStopAndTryAgain)
                     got_some = true;
                 }
                 if (result != Ftp::Result::Next) {
-                    LogDebug("Got result: {}", result);
+                    LogDebug("Got result: {}", static_cast<int>(result));
                     prom.set_value(result);
                 } else {
                     LogDebug(
